improve code

+++ commands/move.go
@@ -1,16 +1,16 @@
 package commands
 
 type Move struct {
-	X       uint32
-	Y       uint32
+	X       int16
+	Y       int16
 	worldId uint32
 	MapId   float64
 }
 
-var WorldIdMax = uint32(2 << 12)
-var MapCoordsMax = uint32(2 << 8)
+var WorldIdMax = 2 << 12
+var MapCoordsMax = 2 << 8
 
-func abs(a uint32) uint32 {
+func abs(a int16) int16 {
 	if a < 0 {
 		return -a
 	}
@@ -18,9 +18,9 @@ func abs(a uint32) uint32 {
 }
 
 func (m *Move) SetFromMapId() {
-	m.worldId = uint32(int64(m.MapId) & 1073479680 >> 18)
-	m.X = uint32(int64(m.MapId) >> 9 & 511)
-	m.Y = uint32(int64(m.MapId) & 511)
+	m.worldId = uint32(uint64(m.MapId) & 1073479680 >> 18)
+	m.X = int16(int64(m.MapId) >> 9 & 511)
+	m.Y = int16(int64(m.MapId) & 511)
 	if (m.X & 256) == 256 {
 		m.X = -(m.X & 255)
 	}
@@ -30,7 +30,7 @@ func (m *Move) SetFromMapId() {
 }
 
 func (m *Move) SetFromCoords() bool {
-	if m.X > MapCoordsMax || m.Y > MapCoordsMax || m.worldId > WorldIdMax {
+	if int(m.X) > MapCoordsMax || int(m.Y) > MapCoordsMax || int(m.worldId) > WorldIdMax {
 		return false
 	}
 	var worldValue = m.worldId & 4095
@@ -42,6 +42,7 @@ func (m *Move) SetFromCoords() bool {
 	if m.Y < 0 {
 		yValue |= 256
 	}
-	m.MapId = float64(worldValue<<18 | (xValue<<9 | yValue))
+	m.MapId = float64(worldValue<<18 | (uint32(xValue)<<9 | uint32(yValue)))
+
 	return true
 }
